backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
Babel==2.8.0
bcrypt==3.2.0
billiard==4.2.1
black==23.12.1
blinker==1.4
boto3==1.34.14
botocore==1.34.162
Brotli==1.1.0
celery==5.3.6
certifi==2020.6.20
cffi==1.17.1
cfgv==3.4.0
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4

configobj==5.0.6
constantly==15.1.0
contourpy==1.3.1
coverage==7.6.10
cryptography==3.4.8
cssselect2==0.7.0
cycler==0.12.1
daphne==4.1.2
dbus-python==1.2.18
distlib==0.3.9
distro==1.7.0
distro-info
Django==5.0.1
django-anymail==10.2
django-appconf==1.0.6
django-cors-headers==4.3.1
django-filter==23.5
django-formtools==2.5.1
django-graphql-jwt==0.4.0
django-guardian==2.4.0
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-storages==1.14.2
django-two-factor-auth==1.15.5
django-webauthn==0.0.2
djangorestframework==3.14.0
djangorestframework-gis==1.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
exceptiongroup==1.2.2
factory-boy==3.3.0
Faker==22.2.0
filelock==3.17.0
flake8==7.0.0
fonttools==4.55.3
GDAL==3.4.1
geographiclib==2.0
geopy==2.4.1
graphene==3.4.3
graphene-django==3.2.0
graphql-core==3.2.6
graphql-relay==3.2.0
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
identify==2.6.6
idna==3.3
importlib-metadata==4.6.4
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jeepney==0.7.1
Jinja2==3.0.3
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.8
kombu==5.4.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
matplotlib==3.10.0
mccabe==0.7.0
more-itertools==8.10.0
msgpack==1.1.0
mypy==1.8.0
mypy-extensions==1.0.0
netifaces==0.11.0
nodeenv==1.9.1
numpy==2.2.2
oauthlib==3.2.0
packaging==24.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.6.0
promise==2.3
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.11.1
pycparser==2.22
pycurl==7.44.1
pydyf==0.11.0
pyflakes==3.2.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyphen==0.17.0
pypng==0.20220715.0
pyrsistent==0.18.1
pyserial==3.5
pytest==7.4.4
pytest-cov==6.0.0
pytest-django==4.7.0
#python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
#PyYAML==5.4.1
qrcode==7.4.2
redis==5.0.1
reportlab==4.0.9
requests==2.31.0
s3transfer==0.10.4
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sqlparse==0.5.3
stripe==11.4.1
systemd-python==234
tabulate==0.9.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
#ubuntu-pro-client==8001
#ufw==0.36.1
#unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
vine==5.1.0
virtualenv==20.29.1
wadllib==1.3.6
wcwidth==0.2.13
weasyprint==60.2
webencodings==0.5.1
XlsxWriter==3.1.9
zipp==1.0.0
zope.interface==5.4.0
zopfli==0.2.3.post1
